Add files via upload
When the stimulus type is freq (the letter with 80% probability) and rare (the number with 20% probability), the average reaction time and average accuracy of the patient group and the control group are calculated respectively. Through the visualization of the data, it is found that the patient group and the control group have no significant difference in response time and accuracy under the same stimulation conditions. However, when freq is used as the stimulus condition, in both the patient group and the control group, better performance, shorter reaction time and higher accuracy than rare as the stimulus condition are obtained. In order to further explore the impact of specific stimulus types (such as a, b, c or 1, 2, 3) on the reaction time and accuracy of the two groups of subjects, further calculations were performed on the data. Through visualization, it was found that the specific stimulus type did not bring significant difference in response time and accuracy.  

When the number of stimuli continued to increase, the accuracy and reaction time of the control group and the patient group were basically the same. In addition, in the two groups of experimental groups, the correct rate under Freq stimulation conditions was higher than Rare's stimulation conditions from the beginning, and the reaction time under Freq stimulation conditions was shorter from the beginning than when Rare was the stimulation conditions. This shows that the increase in the number of stimulations can enable the control group and the patient group to obtain nearly the same task performance when receiving the same stimulation conditions. In addition, the two stimulus conditions with different probabilities will show a stable difference in response time and accuracy as the number of stimuli increases.